ENH: Support alternate dim orders in image_from_xarray
The order of dims can be different from the result of xarray_from_image.
If they are not in the required ordere, the data is re-organized.

image_view_from_array defaults to a return type of VectorImage for
non-UC, 3 or 4 component images to support more pixel component values.
ImageDuplicator is wrapped for the required VectorImage types from
image_from_array.