COMP: Changed conditional include of emmintrin.h to use preprocessor
For compatability with Universal Binaries on macOS, change how we decide if we can #include <emmintrin.h>.

Using a trycompile to check if #include <emmintrin.h> compiles doesn't work for Universal Binaries becuase the check will only be done for the native architecture, not both architectures (where indeed there would be the opposite answer).

The mess of checks is necessary to support various compilers.