COMP: removed hardcoded `-march=corei7` compilation flag
This fixes building for x86_64 from an arm64 Mac.

In my view, a library like ITK should not be assuming what kinds of CPU the resulting executable will be run on. It can't know if the result will only be run on the building machine, or deployed to customers running all kinds of models of computer. Only the programmer building ITK knows this.