STYLE: Replace "itk::ERROR" by "ITK ERROR" in description of exception
Previously, the description text of an exception thrown by
`itkSpecializedMessageExceptionMacro` would start with the text
"itk::ERROR", which seemed to suggest that it involved an `itk::ERROR`
object, function, or type. However, there is no such `ERROR` in the
`itk` namespace!

It appears clearer to use the text "ITK ERROR" instead, to indicate
that the error involves the ITK library.