WIP: BUG: Replace dynamic_cast by Convert, doing implicit type conversion
If possible, it is preferable not to use of `dynamic_cast` because of
performance costs, and sometimes segmentation faults (as observed on
MacOS).

`Conversion::Convert<T*>(ptr)` safely converts a pointer to `T*`. It
returns null if `ptr` is not implicitly convertible to `T*`.

Calls to this new conversion function replace `dynamic_cast` in 18 places.