BUG: Fix regression in PolygonSpatialObject::IsInsideInObjectSpace
This is a followup to the bugfix for issue #1082 where there was a
regression in the code on determining the inside/outsideness of points.

There are some cases where the algorithm still fails as there are two
boundary checks in the code which should be identical.
1. Test on adjacent segment pairs from start to finish
2. Test on segment pair from finish point to start point on closed
   polygons.

The previous bugfix fixed the code in item 1, but did not update the
check (which should be identical in item 2).

This should address this issue making both boundary on segments the same.