COMP: fix link error building Universal Binary on Intel Mac
Without this, building Universal Binary on an Intel Mac (but not on an ARM Mac) gives link errors for undefined symbols like _itk_png_do_expand_palette_rgb8_neon and a few others.

Introspecting the CPU type of the build machine is fundamentally incompatible with the idea of Universal Binaries.