ENH: StatisticsLabelMapFilter use improve integer histogram
For 8 and 16-bit integer set the number of histogram bins to the
number of values. Also add bin margins so that the center of the bins
exactly match the integer values.

Fix bin selected when there are an odd number of elements. Now
correctly choosing the middle bin.

This combination results in exact median computation for 8-bit and
16-bit integers.