WIP: ENH: Added type hints for wrapped Python classes and methods.
This is a method of generating IDE type hints for the dynamically generated ITK Python classes
and dictionaries. The type hints are generated through .pyi files in the itk-stubs directory.

In this approach, the stub files tell the IDE that the dynamically generated itkTemplate (i.e.
itk.ListObject) objects used to specify types for classes are actually classes of that name. This
allows type hints to be used.

Two methods are typed so that the `New()` method appears to return a generic ‘Typed’ object
for that class and the dictionary class access appears to return the same class for which the
`New()` method can be used.

The `New()` appears to return a generic `Typed` class with the name
`Typed<class_name>PROXY`. This class then defines generic type hints for all methods of the
class. This way the user has an idea of what methods are available and parameters are
needed, but may not know the exact type or dimension for some parameters or returned values.
This class is not real and is only used to generically describe type hints for all methods of a
class.

The current code is a proof of concept using `itk::Statistics::ListSample` as an example. The
methods and parameters for this class and its parent classes are typed.

--UPDATE--

The `TypehintGeneration.py` file has been added to help assist in the typing process. This file
has several functions that parse the help text for a class and generate useful python pyi code
formatted as described above. The Doxygen documentation is parsed and used to add
descriptions to itk class methods when available.

While this parser works on the majority of classes. There are currently three problem areas:
1. It can only generate complete code for instantiatable classes. However, inherited method hints can be displayed for a class.
2. The c++ to python type conversion is not complete. Ideally, there would be a better method for doing this as my current approach isn't ideal. The typed itk classes also make this part difficult because they can end in many different ways. This is currently being worked around by a large block of auto-generated code but does not fix all issues.
3. Typed PROXY class hints have Unions for parameters with varying types. Currently, the types are gathered by running the parser for each type variation of the class. While this works, it's quite slow.

Hints for the `itk.Vector` class have been generated and added to the `__init__.pyi` file. Note
that most C types still don’t have conversions so they are not yet converted.