DOC: Fix description of SpatialObjectToImage registration
Corrects the description of the motivation for SpatialObject
registration:

In the particular case of ImageToSpatialObject registration, the
Transform to be computed is the one mapping points from the SpatialObject
into the Image.  This allows the SpatialObject to drive the location and
geometry of the measurements made in the image - thereby if the spatial
object is sparse and/or contains varying geometric features, the metric
applied to measure how well that object matches with the image can be
rapidly computed only at the sparse locations and using measures that
match the local geometry that should be at those locations in the image.
This is particularly useful for fast intra-operative registration when the
pre-operative data is used to define a SpatialObject and can anticipate
how that object will appear in the intra-operative images.