BUG: Must add a point to a spatial object before requesting world info
A spatial object point has a position in object space, but until
it is added to a spatial object (which potentially exists in a
hierarchy of objects with parent/child transforms) the point
does not have a position in world space.  So, the test has been
updated to add a point to an object before asking for that point's
world space position.