ENH: Simplify a directed Hausdorff distance test
The test
itkDirectedHausdorffDistanceImageFilterTest2
looked like it does a lot more than what it actually does.
It took two input images, but it's meant to be used with identical
input images (or ones where the support of one image is a subset
of the support of the other). It needlessly took a boolean parameter
that it doesn't use for testing the logic of the filter.

The interesting aspects of directed Hausdorff distance are
pretty much already tested in
itkDirectedHausdorffDistanceImageFilterTest1.cxx
now, so it's okay for this test to cover other trivial aspects.

See comments at the end of issue #2445
concerning
itkDirectedHausdorffDistanceImageFilterTest.cxx,
which was the previous name of the test in question here.