BUG: Use GenerateOutputInformation and pipeline isolation
The TransformGeometryImageFilter::GenerateData method was refactored
for correct pipeline behavior. The output Image's meta-data is
correctly modifed in the GenerateOutputInformation method, so that
down stream filters can uses the information. The composite pipeline
is not isolated from the inputs and output using the double graft
convention. However, only the bulk data pointer and buffered region
are copied to the output, as not to change the geometry information
previously set.