ENH: Register/UnRegister only in SmartPointer Con/Destructor.
Only in the constructors and destructors for the SmartPointer should
we invoke the underlying object's Register() or UnRegister() methods
to handle the smart pointer reference count. The original code had
m_ReferenceCount initialized to 1 in the construction of an underlying
object (despite that there were 0 smart pointers) and had multiple
functions responsible for UnRegister() of that extra count after the
created object was assigned to a SmartPointer.

Note that Register() and UnRegister() are okay and have been kept in
functions that act as if they were SmartPointer constructors or
destructors, such as those that pass raw C++ pointers to be used by
Python code. Similarly they are kept in ObjectFactoryBase because, for
backward API compatibility, lists of ObjectFactoryBase * are used
instead of lists of ObjectFactoryBase::Pointer; and Register() and
UnRegister() are used manually each time a pointer is inserted or
removed from a list.