COMP: Initialize CXX standard in ITK_USE_FILE
Code that builds against ITK may not be built with a sufficient or
compatible C++ standard.

Use ITKInitializeCXXStandard.cmake in client code, which calls

  include(${ITK_USE_FILE})

according to our existing documentation, which will set
CMAKE_CXX_STANDARD and CMAKE_CXX_STANDARD_REQUIRED if not already set.

To address build errors as discussed in:

  https://discourse.itk.org/t/hello-world-from-itk-examples-installation-not-compiling-for-itk-5-3/4441/8