BUG: Fix linking by making LBFGS2Optimizerv4 a template
Having LGBFGS2 being a compiled class in the Optimizerv4 library
created linking errors from duplicate symbols on Windows with shared
libraries in certain situations. The cause is that the parent classes
of LBGFS2 were templated, and their link specifcation is not
compatibile for them being in a shared library.

The implemented solution was to make the LBFGS2 Optimizer a template
class like its parents so that its symbols will not be in a shared
library.