ENH: Remove LBFGS2's pimpl member
The lbfgs library is not part of the public interface for the
Optimizerv4 library since the LBFGS2 optimizer is no longer compiled
into the library. Since the lbfgs methods and structures are used in
the hxx files, there is not purpase to keep the structures and usage
encapsulated in a private implementation.