ENH: Adding Gaussian Curvature Calculator in ITK Mesh with tests
Gaussian curvature is needed for Thin Shell Demons based Mesh registration.
Calculator design pattern is followed so that only the calculated curavature data needs to be stored.
Filter pattern is not suitable for this purpose as entirely new mesh needs to be created.

Curvature data is stored in a Double Vector Container. Mesh PointData Container is not suitable for
this purpose as the Input Mesh can have custom Data Container.

Min and Max curvature require Mean curvature and they will be added in future.
To support them TriangleMeshCurvatureCalculatorEnums has been introduced in the current implementation.

Python wrapping for Gauss curvature also added.