ENH: Add test for inverse Jacobian of AffineTransform.
The function ComputeInverseJacobianWithRespectToPosition is now tested to provide the inverse of the transformation matrix of the AffineTransform. This function was not tested before.

It has been motivated by the BUG correction of this function in its superclass = MatrixOffsetTransformBase. The version previous to this BUG correction should not pass the test.