BUG: ComputeInverseJacobianWithRespectToPosition didn't do inversion
itk::MatrixOffsetTransformBase::ComputeInverseJacobianWithRespectToPosition was returning directly the matrix (the same as ComputeJacobianWithRespectToPosition). It has been corrected to return the inverse matrix.

This correction has required also the correction of a few function return and parameters' Types.
Those corrected Types were inconsistent when instantiated with InputImage and OutputImage of different dimensions.

Apparently the bug was inadvertently introduced in
61e2be7e6dd666ccf6a7db2f6175bef310381927 (PERF: Improve MatrixOffset Jacobian computation performance)
and subsequently not fixed but got mixed wrong JacobianTypes in
212cae522d8451ea089c41f8a151279e1dd17042 (PERF: Change type of Jacobian w.r. position to vnl_matrix_fixed).