ENH: Adding Test for Array PixelType in Mesh Wrapping
Added test in Python to demonstrate the usage of Array PixelType in Mesh.
Usage of Array allows to change the size of pixeldata as per requirements later.
Using Array, however breaks the API support for SetElement and SetPointData in PointSet.
The only way to alter the data is through reference and for that CreateElementAt method is
used in the VectorContainer.