ENH: TimeBounds' normalization made optional
Added boolean flag to interpret LowerTimeBound and UpperTimeBound as rates
in the velocity field time span. This is equivalent to consider that the
velocity field is defined in the normalized time interval [0, 1], ignoring
the input origin and spacing in the temporal dimension.

The default is true for backwards compatibility.

ENH: TimeBounds' normalization made optional

Added boolean flag to interpret LowerTimeBound and UpperTimeBound as rates
in the velocity field time span. This is equivalent to consider that the
velocity field is defined in the normalized time interval [0, 1], ignoring
the input origin and spacing in the temporal dimension.

The default is true for backwards compatibility.