ENH: Add boolean macros to `itk::SpatialObjectWriter` boolean ivars
Add boolean macros to `itk::SpatialObjectWriter` boolean ivars.

Exercise the boolean macros and the boolean ivar Set/Get methods.

Take advantage of the commit to simplify and improve the style and
readability around such testing: use the `ITK_TRY_EXPECT_NO_EXCEPTION`
macro to avoid boilerplate code when updating a filter that can trigger
exceptions, and take out the rest of the statements of the `try/catch`
block.