BUG: Avoid duplicate progress reporting in general case
Before 840bae71744bbac577e4802e4e42863ca2f1e90e
for some filters progress would go from 0% to 100% twice,
overflowing from 100% to 0% about halfway.
After 840bae71744bbac577e4802e4e42863ca2f1e90e,
the progress would fill up to 100% during first half of the work,
and stayed there the entire second half of the work.

With this commit, progress is reported as expected:
it fills up to 50% during first half of the work,
and fills up to 100% during second half of the work.
The progress should be filling up smoothly, in small increments.

This is because we now disable UpdateProgres in  MultiThreader
in class constructors of classes which report filter progress.

Closes #2875.