STYLE: ConstNeighborhoodIterator in-class default member initializers
Added in-class default member initializers to data members of
`ConstNeighborhoodIterator` and `ConstNeighborhoodIteratorWithOnlyIndex`.

Removed the apparently redundant `ResetBoundaryCondition()` call from
the default-constructor of `ConstNeighborhoodIterator`.

Defaulted (`= default`) the default-constructor of
`ConstNeighborhoodIteratorWithOnlyIndex`.

In accordance with C++ Core Guidelines, January 3, 2022: "Prefer
in-class initializers to member initializers in constructors for
constant initializers"
https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rc-in-class-initializer

Related Clang-Tidy check:
https://clang.llvm.org/extra/clang-tidy/checks/modernize-use-default-member-init.html