ENH: Add macro to selectively catch factory instantiation failures
Adds `itkFactoryOnlyNewMacro` as an alternative to `itkNewMacro` and
`itkFactorylessMacro` for defining the `New()` method for constructing
object smart pointers.

The ITK object factory can be used for overriding class constructors and
ipmlicitly returning specialized constructs. `itkNewMacro(Self)` is
commonly used to define a `New()` method for instantiating objects by
first checking for factory methods and then relying on simple
instantiation with the `new` keyword if (when) no factory is found. This
is suitable for most classes which define their own concrete
implementation; however, FFT filters specifically are structured such
that the base class does not fully define a backend and should not be
constructed on its own.

The current method of returning `nullptr` if no suitable FFT factory is
found is not common practice in ITK and results in segfaults when we try
to query methods on the returned instance, which can be difficult to
debug. ITK attempts to register factories for Vnl FFT implementations by
default, so the absence of a factory indicates a failure condition.
Rather than requiring calling methods to explicitly check for `nullptr`
it is more reasonable to throw an exception to signify that expected
factories are absent.

With these changes, attempting to instantiate an FFT filter for which no
factory is available returns an ITK RuntimeError giving the line number
of the failed instantiation call and a message explicitly stating the
reason for failure.

Note that `itkFactoryOnlyNewMacro` also adds `CreateAnother` and `Clone`
methods to base FFT classes, which were missing.