ENH: Expose kernel parameters in `DiscreteGaussianImageFilter` interface
`itk::DiscreteGaussianImageFilter` exposes parameters for generating a
Gaussian kernel to error and size specifications, but previously did
not make information on the kernel that was actually generated available
to other classes.

These changes expose protected methods to allow subclasses to create
kernels to the same specifications and also publicly exposes information
on the kernel sigma when image spacing is being accounted for.