ENH: deprecate INVALID_HANDLE_VALUE
It was only used in one place, and the stated reason for its use no longer seems true. Perhaps its a relic of some previous refactoring.

The superficial reason to deprecate it though is that on some platforms it's an inteeger and on some it's a pointer. This means you need to initialize it to zero, but then platforms where it should be a pointer pedantically suggest to use nullptr.  As this is literally the last such "use nullptr" warning, it's nice to fix it.