BUG: Fixed import issues with Python stub interface files.
Previously, some wrapped classes inherited from classes defined in a seperate file.
These classes were not imported. This fix allows the classes to be imported
without requiring each file to know the name of the file that the
class is located in. All proxies are imported in the _proxies.pyi file
which can then be imported by different template classes. The Proxy
and Template classes were split to allow the leading underscore to
be removed from the proxies without causing extra 'fake' classes
to appear for the user.