BUG: GDCMImageIO accepted images with non-orthogonal direction cosine.
The DICOM standard requires that "The row and column direction cosine
vectors shall be orthogonal, i.e., their dot product shall be
zero. The row and column direction cosine vectors shall be normal,
i.e., the dot product of each direction cosine vector with itself
shall be unity."

ITK images support arbitrary affine matrices as the direction
cosine. When they are not compliant with the DICOM requirements GDCM
write silently changes the direction cosine to be compliant. This
commit enforces the requirement in ITK so the user can choose if they want to
explicitly modify the matrix to comply with the DICOM requirements or
possibly write using a different format which supports affine matrices
for direction cosine.