ENH: Remove support for wrapping with RUBY
The RUBY bindings are not tested.  This batch of untested
code makes refactoring the wrapping process more difficult
as it requires changes to code that is very unlikely to be
used.

NOTE: parts of the JAVA, PERL, & RUBY wrapping support has
been commented out since a17865c14e6 2018-07-07, and does
not work for wrapping currently.