COMP: Simply prefix based on internal/external
The global value set in the macro
`itk_wrap_modules_set_prefix` was a constant
value known at configure time based solely on
the status of being an internal or external
module build.

By placing the conditional setting of the global
variable at the begining of the commmonly
used code, we can avoid making many
calls that redundantly set the value
to the same value.