ENH: Check GTest test method return value
Check GTest test method return value.

GTest ignores return values from test methods, hence, the return values
of the methods that are being called would be similarly ignored if such
methods are intended to provide the return value of a test. This commit
saves the return values of the called methods, and compares them to a
successful test outcome value.