STYLE: Use MakeFilled, () or {}, instead of `Vector(const ValueType &)`
A call to `MakeFilled<VectorType>(value)` expresses more clearly than
`Vector(const ValueType &)` that an _entire_ Vector is to be filled by
the specific scalar value.

Zero-initialization of a Vector is achieved in the most generic way, by
an empty pair of parentheses `()` or an empty pair of braces `{}`,
rather than by passing 0.0 to the explicit `Vector(const ValueType &)`
constructor.