COMP: Do not use ITK_AUTOLOAD_PATH when ITK_WRAP_PYTHON is set
For the use case of Slicer using ITK Python wheels, ITK_AUTOLOAD_PATH
is set, and the Python wheel binaries will load the Slicer ITK binaries,
which will cause a crash.

Dynamic loading in Python is unnecessary and can be performed in a more
explicit and controlled manner by Python scripting, so disable this
feature by default when ITK_WRAP_PYTHON is set.