BUG: Set number of targets for FMUG image filter
Set the number of targets ivar to the appropriate value for the
`itk::FastMarchingUpwindGradientImageFilter` class.

The existing implementation was not setting the `m_NumberOfTargets` ivar
to the appropriate values depending on the `m_TargetReachedMode` value.

Ensure that the a sufficien number of target points exists for the
desired `m_TargetReachedMode`.

Exercise and test all possibilities.