DOC: Document methods in header file
Document methods in header file: Doxygen picks the method documentation
from the header files, and leaves the method documentation empty if it
does not find it there.

Take advantage of the commit to fix the method descriptions where
necessary (i.e. provide an accurate description of the method, do not
use the description of another method, etc.), and improve the wording
where appropriate (avoiding expressions like "This method (...)"),
including fixing typos and grammar.

Remove uninformative comments.