BUG: Fix largest possible output region in `itk::ConvolutionImageFilter`
As a result of disconnecting input for its internal mini-pipeline
`itk::ConvolutionImageFilter` previously set the largest possible region
for its output image to be the same as the output buffered region.

This fix updates output information prior to grafting mini-pipeline
results onto the filter output so that the largest possible output
region reflects the size of available input data with possible
adjustment for the filter's region mode. Critical for image streaming
operations.