PERF: Reduce verbosity of `itk::GaussianOperator` warning
`itk::GaussianOperator` can be set with a `MaximumKernelWidth` parameter
such that a kernel that would otherwise overflow this size is cropped to
the maximum width. Cropping to the set maximum width happens frequently
in cases such as Gaussian blurring for multiscale pyramid generation and
is expected behavior, but the previous warning message makes it seem
like a failure case. This commit reduces the verbosity of the message
from a warning to a debug printout so that developers can get insights
into kernel cropping when needed, but are not alarmed by default warning
printouts.