BUG: Use absolute module path PyCapsule_Import
Addresses:

```
ImportError: PyCapsule_Import could not import module "_ITKCommonPython"
```

That can occur when unpickling for Dask, for example.

According to the `PyCapsule_Import` documentation, internally it calls
`PyImport_ImportModule`. According to the `PyImport_ImportModule`
documentation, this requires the absolute path -- we add the `itk.`
prefix.

In most other contexts, the `_ITKCommonPython` module is likely to
already be imported. Since `_ITKCommonPython` gets loaded via
PyImport_ImportModule, it is not registered with the ITK lazy loading
module system, so we add `from . import _ITKCommon`.