COMP: Allow changing CMAKE_INSTALL_LIBDIR
Many paths are hard-coded to assume that the install
location of libraries is always 'lib'.  When GNUInstallDirs
is used on CentOS the default location of libraries is
sometimes set to 'lib64'.

If a user explicitly sets CMAKE_INSTALL_LIBDIR, then
that location should be respected by all the third-party
interall builds as well.

Previously both FFTW and DCMTK would install their libraries
in 'lib64' but would look for the libraries in the
'${CMAKE_INSTALL_PREFIX}/lib' directory.