BUG: Avoid dangerous use of putenv
https://stackoverflow.com/a/54279521

While std::getenv is part of the C++ standard, the putenv function isn't.

As you can see from the linked POSIX reference for putenv, it's argument
is of type char *.  This is very important, and one thing that differs
between C and C++: In C a literal string can be passed to functions
expecting char *. In C++ all literal strings are constant and can only
be passed to functions expecting const char *.

Use the itksys::Process::PutEnv wrapper that provides
cross platform safe wrappers for setting/getting
the environmental variables.