ENH: Add script for Python3 find_package upgrade
Now that cmake FindPython3 can be universally
used, prefer to only use that set of variable
names for identifying the python executables.

This maintenance script can be used to assist with moving from
the deprecated find_package(PythonInterp) and find_package(PythonLib)
to the preferred find_package(Python3) variable names.