PERF: Using stl stable_sort instead of a custom class and sort
The algorithm requires a stable sort of image indexes based
on pixel intensity. The paper recommends counting sort but
that has its own problems when using bigger than 8 bit pixel
types or floats - i.e. need additional scans of the data
to figure out range etc. The original implementation
of this class used a structure with both position and brightness.
This version abandons that structure and uses an offset array
only, saving some RAM and simplifying the code.

There's an opportunity to activate stl parallel sorting in the
future.