BUG: Set the maximum number of iterations to a non-zero value in test
Set the `m_MaximumNumberOfIterations` `itk::MultiLabelSTAPLEImageFilter`
instance variable to a non-zero value in the corresponding test.

The value was set to 0 in commit e2511e7c, following the default value
of the instance variable. However, this made it such that when updating
the filter no update was being actually performed due to the condition
in
https://github.com/InsightSoftwareConsortium/ITK/blob/e2511e7c0f0636439f08adc6b81abba18eceec33/Modules/Segmentation/LabelVoting/include/itkMultiLabelSTAPLEImageFilter.hxx#L249

This also made the coverage decrease. Setting the variable to a non-zero
value restores the original test behavior and provides the same
confusion matrix results.