DOC: Document methods in header file
Document methods in header file: Doxygen picks the method documentation
from the header files, and leaves the method documentation empty if it
does not find it there.

Increase consistency in the characters used for documentation blocks,
and follow the ITK SW Guideline.

Take advantage of the commit to fix the method descriptions where
necessary (i.e. provide an accurate description of the method, do not
use the description of another method, etc.), and improve the wording
where appropriate (avoiding expressions like "This method (...)"),
including fixing typos and grammar.

Finish method documentation sentences with a period.

Remove the documentation of trivial methods: constructors, destructors
and `PrintSelf` methods.