DOC: Fix unexpected token in class documentation
Fix unexpected token in the `itk::TransformGeometryImageFilter` class
documentation.

Remove the `\param` Doxygen commands from a non-function documentation
block, and reuse part of the documentation to describe what the filter
yields as its output.

Avoid duplicating the type of transformation the filter accepts in the
corresponding `Transform` macro (as it is already present in the class
documentation).

Take advantage of the commit to remove an unnecessary/uninformative part
from the `InputImage` macro.

Fixes:
```
Modules/Core/Transform/include/itkTransformGeometryImageFilter.h:45:
warning:
unexpected token in comment block while parsing the argument of command param
```

Reported for example in:
https://open.cdash.org/viewBuildError.php?type=1&buildid=8167118